+++ webhook-service/src/server.ts
@@ -6,11 +6,12 @@ import { metricsRoutes } from "./routes/metricsRoutes";
 import { dlqRoutes } from "./routes/dlqRoutes";
 import userRetryRoutes from "./routes/userRetryRoutes";
 import { errorHandler, requestLogger } from "./middleware";
-import { ipValidator } from "./middleware/ipValidator";
+import { ipValidator } from "./middleware/mpesaIpMiddleware";
 import db from "./drizzle/db";
 import { sql } from "drizzle-orm";
 import authRouter from "./routes/auth.routes";
 import settingsRoutes from "./routes/settings.routes";
+import 'dotenv/config';
 
 const app = new Hono();
